Prepare for possible import of databases
In order to be able to import into a central database add 2 fields
to local database to be able to mark imported experiments.

"imported" mark if imported or not.
"uuid" is used to avoid importing twice in case we already
imported but we fail to mark "imported" field locally, sequence will
be:
- try to import "imported" == 0
- if "uuid" is already present in destination mark "imported" to 1
  and finish
- copy the experiment to central db and commit the remote transaction
- mark "imported" to 1 and commit locally