ui.web.graph: plot Keyframe sizes and periods in the Viewer
These modifiers 1) highlight the keyframes sizes and 2) compute the
keyframe period.

The algorithm works by marking the biggest frame as key frame, then
the second biggest, etc. It relies on two parameters:

    MIN_KEYFRAME_PERIOD = 11

the minimum number of frames between two keyframes. If the biggest
frame is closer than that to a keyframe, then it's just a "big" normal
frame.

    KEY_NORMAL_SIZE_RATIO = 33/100

indicates how much bigger a frame must be to be a possible
keyframe. When the biggest frame size drops below the ratio, no more
keyframes can be found.