ui.web: rewrite Matrix scripting/viewer to improve scripted properties
It now works this way:

    scripted_properties:
      webpage:
        cubemap: https://webglsamples.org/dynamic-cubemap/dynamic-cubemap.html
        still: https://www.redhat.com/

      resolution: 1920x1080, 1366x768, 848x480

    scripts:
      webpage:
        before:
          - vm do killall firefox 2 > /dev/null
          - vm open $webpage
          - sleep 20
        after:
          - vm do killall firefox 2 > /dev/null

      resolution:
        before:
          - vm resolution $resolution
        after:
          - vm resolution 1200x1920

`webpage` is a named property (name -> actual value)
`resolution` is a list-property (value1, value2, ...)

For each of the scripted properties, the `before` and `after` scripts
will be executed when the value should change (eg, (after(previous),
before(new))).

Scripted properties are now first class variables, same as the
`matrix` properties (for encoder parameters).