ui.web.matrix_view: reorganize dcc.Graph objects to speed up the plotting
We used to have one dcc.Graph per 'Stats', that was plotted when the
stat was selected, or not displayed when it wasn't selected. That was
good when few stats were available, but then it was slowing everything
down.

Now, there are NB_GRAPHS (=3) dcc.Graph objects, that will plot the
1st/2nd/3rd stats in the list (the rest is ignored).